% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutated.R
\name{mutated}
\alias{mutated}
\title{Mutate}
\usage{
mutated(population, prob)
}
\arguments{
\item{population}{parents population}

\item{prob}{probability to mutate}
}
\value{
Mutated data
}
\description{
Mutate the generated data
}
\examples{
mutated(matrix(rep(1,20), nrow = 2), 0.5)
}
