% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankselection.R
\name{rankselection}
\alias{rankselection}
\title{Rank Selection}
\usage{
rankselection(scores)
}
\arguments{
\item{scores}{parents scores}
}
\value{
A vector of parents index of generated data
}
\description{
Sample from scores by rank selection method and get index
}
\examples{
score <- runif(2)
rankselection(score)#get 4 samples from score
}
