% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tournament.R
\name{tournament}
\alias{tournament}
\title{Tournament Selection}
\usage{
tournament(scores, k = 2)
}
\arguments{
\item{scores}{The scores of model fitness}

\item{k}{The tournament size, default is 2}
}
\value{
An array of population indices
}
\description{
Sample from population
}
\examples{
scores <- 1:10
tournament(scores)
}
